review - refactoring inner class to be top level class. Adding javadoc
